Adds margin of safety and current market price.
This is the MVP of the product, really. There's an awful lot on the
backend that's lacking, but it produces something approximating the
manual process I'm used to.

The script now pulls data for the various tickers on the watchlist, or
supplied directly. It calculates some of the key compound annual growth
rates, and pulls out some important ratios.

The income and balance sheet data I currently care about gets pulled
into a dataframe; this is written to an excel, along with the ratios,
compound growth rates, and margin of safety and current market price.

Herein lies the first set of changes required; a lot of the positioning
of data is hardcoded and would fail if I wanted to add more information,
for example. The formatting is hardcoded, both in terms of style and
location. All of this should instead be data driven to make it easy to
change.

Another next step is going to be to clean up all the code; at the moment
it's just one large, messy, unmaintainable morass. Functions, other
modules, etc., are sorely lacking.